requirements.txt
numpy==1.20.1
opencv-python==4.5.1.48
Pillow==8.1.1
svgwrite==1.4.1
